--- a/crates/subspace-proof-of-space/src/chiapos.rs
+++ b/crates/subspace-proof-of-space/src/chiapos.rs
@@ -8,11 +8,7 @@ mod tests;
 mod utils;
 
 pub use crate::chiapos::table::TablesCache;
-use crate::chiapos::table::{
-    fn_hashing_input_bytes, metadata_size_bytes, x_size_bytes, y_size_bytes,
-};
 use crate::chiapos::tables::TablesGeneric;
-use crate::chiapos::utils::EvaluatableUsize;
 
 type Seed = [u8; 32];
 type Challenge = [u8; 32];
@@ -20,18 +16,7 @@ type Quality = [u8; 32];
 
 /// Collection of Chia tables
 #[derive(Debug)]
-pub struct Tables<const K: u8>(TablesGeneric<K>)
-where
-    EvaluatableUsize<{ x_size_bytes(K) }>: Sized,
-    EvaluatableUsize<{ y_size_bytes(K) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 1) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 2) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 3) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 4) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 5) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 6) }>: Sized,
-    EvaluatableUsize<{ metadata_size_bytes(K, 7) }>: Sized,
-    EvaluatableUsize<{ fn_hashing_input_bytes(K) }>: Sized;
+pub struct Tables<const K: u8>(TablesGeneric<K>);
 
 macro_rules! impl_any {
     ($($k: expr$(,)? )*) => {
@@ -93,13 +78,5 @@ impl Tables<$k> {
     }
 }
 
-// These are all `K` which can be safely used on 32-bit platform
-#[cfg(feature = "all-chia-k")]
-impl_any!(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 19, 20);
-impl_any!(16, 17, 18);
-
-// These are all `K` which require 64-bit platform
-#[cfg(target_pointer_width = "64")]
-impl_any!(32, 33, 34, 35);
-#[cfg(all(target_pointer_width = "64", feature = "all-chia-k"))]
-impl_any!(21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 36, 37, 38, 39, 40);
+// Only these k values are supported by current implementation
+impl_any!(15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25);