--- a/crates/subspace-proof-of-space/src/lib.rs
+++ b/crates/subspace-proof-of-space/src/lib.rs
@@ -9,7 +9,6 @@
     const_trait_impl,
     generic_const_exprs,
     int_roundings,
-    iter_collect_into,
     portable_simd,
     step_trait
 )]