These are improvements targeting https://github.com/subspace/subspace/issues/1502

There was one bug in old implementation that was reproducible on k22 (and caught by debug assertions), I decided to not fix it since old implementation will be gone.

In this version I have deleted generic implementation and replaced with implementation that only supports k15 to k25 and uses numeric data structures in most cases. Portable SIMD instructions were used for first table construction and resulted in significant wins for that small section of work.

Before:
```
chia/table/single       time:   [194.73 ms 194.86 ms 195.04 ms]
chia/table/parallel     time:   [29.027 ms 29.235 ms 29.496 ms]
chia/quality/no-solution
                        time:   [33.519 ns 35.358 ns 36.038 ns]
chia/quality/solution   time:   [89.765 ns 89.773 ns 89.782 ns]
chia/proof              time:   [462.50 ns 462.82 ns 463.10 ns]
chia/verification       time:   [12.021 µs 12.029 µs 12.044 µs]
```

After:
```
chia/table/single       time:   [133.30 ms 133.45 ms 133.66 ms]
chia/table/parallel     time:   [22.230 ms 22.385 ms 22.538 ms]
chia/quality/no-solution
                        time:   [20.748 ns 20.749 ns 20.750 ns]
chia/quality/solution   time:   [43.774 ns 43.791 ns 43.811 ns]
chia/proof              time:   [183.42 ns 183.68 ns 183.89 ns]
chia/verification       time:   [11.574 µs 11.619 µs 11.661 µs]
```

Those numbers are for k17. For k22 improvements are also decent, but seem to be smaller in percentages (still double digits, but it takes a long time to re-run them).

Overall on my machine I can create Chia table for k22 in under 800ms when using all cores.

Still need to figure out how to apply SIMD and better utilize parallelism in tables 2..7, also there might be opportunity for instruction pipelining, but I have not found good leads for those yet.

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/subspace/subspace/blob/main/CONTRIBUTING.md)